# C2 does not do constant-folding of xor