C2 does not eliminate XOR nodes when it's arguments are constants. This has a noticeable effect on `Long.expand` on architectures that don't have instructions equivalent `PDEP` to be used in an intrinsic.

This patch demonstrates a potential fix to the problem, but there might well be better ways to do it.