@j3graham 2025-01-13
Sample program:
```
public class XorTest {
    public static void main(String[] args) {
        long t=0;
        for (int i = 0; i < 10_000; i++) {
            t+=doXor();
        }
        System.out.println("t = " + t);
    }

    static long doXor(){
        long c=42;
        return c ^ 2025L;
    }
}
```

IGV before patch (constant nodes are red)

![Xor before](https://github.com/user-attachments/assets/8cecb845-8273-4647-91fa-919dec02a4f0)


IGV after patch
![25 Return](https://github.com/user-attachments/assets/36f02474-df70-4a0c-b1c0-71c02732501f)