@SirYwell 2025-01-14
The code for this optimization seems to be present already in the `add_ring` methods, and from my understanding those should be called by `AddNode::Value`. Does that mean `XorINode::Value` and `XorLNode::Value` never call that method in your tests, and instead always end up in the `if` directly below your changes? Maybe that `if` should rather check to ignore constant values?