@eme64 2025-01-22
@j3graham Since this is a performance improvement: do you have any benchmark that shows a speedup?