@j3graham 2025-01-22
Some rough performance numbers with existing tests:

Aarch64 M4 Pro
Baseline:
```
Benchmark       (size)  Mode  Cnt  Score   Error  Units
Longs.compress     500  avgt    5  0.512 ? 0.002  us/op
Longs.expand       500  avgt    5  0.503 ? 0.014  us/op
```

This PR:
```
Benchmark       (size)  Mode  Cnt  Score   Error  Units
Longs.compress     500  avgt    5  0.420 ? 0.008  us/op
Longs.expand       500  avgt    5  0.312 ? 0.002  us/op
```

It's sensitive to what the mask value is. With the mask from https://github.com/openjdk/jdk/pull/22928, the numbers are:
Baseline:
`Longs.expand2      500  avgt   15  0.494 ? 0.002  us/op`

This PR:
`Longs.expand2      500  avgt    5  0.242 ? 0.002  us/op
`