@j3graham 2025-01-23
Remaining issues are
- do we keep the original code that calculates the the bounds, or substitute something that looks easier to explain.
- do we omit the check for r0->_hi and r1->_h1 being > 0. Without it r0=(0,0) and r1=(0,1) , gets past the check that both are constant with r0->_hi equal to zero.  In the original code this would cause a problem - except that 0^x should have been simplified already, so it would be prevented. But that code is a little further away, so not immediately clear.