@j3graham 2025-01-25
I have added tests specifically for the power of 2 case, and updated `add_ring` to use the simpler logic. I've also added some examples to help clarify it.

After https://github.com/openjdk/jdk/pull/17508, this might lend itself to an even more intuitive solution.