@eme64 2025-01-27
@merykitty @j3graham 
> I would advise against adding more complicated jtreg tests as reasoning about them can be tricky. For these operations I think it would be more preferable to use a gtest which can directly inspect the details of Type objects while jtreg tests can only reason about them indirectly. You can generate some pairs of Types, calculate their xor, then, try to sample some values inside the operands and assert that the results would be a member of the resulted Type.

Yes, actually `gtest` may be a better solution if that works!