@j3graham 2025-01-28
I'm going to need some guidance on using gtest for this. I can't really call xor on the types - that would need a Node which doesn't seem to easily be created in a stand-alone way. I can move the tricky code from add_ring to a static method to test that more directly. I'm having trouble getting a test to run without the vm crashing.

As soon as I  try creating a type, for example `const TypeLong *v=TypeLong::make(0)`, I get something like this:

`