@eme64 2025-01-30
That is unfortunate indeed. It would be nice if we had some way of running the Types independently of the CompilerThread. It would really improve our ability to test Types.