@j3graham 2025-01-30
By narrowing down the scope of my testable static method, I've now managed to write some gttests that exercise the most mysterious part of the code. I've also built up more in the IR tests - probably to the point where they have become too complicated, and have significant overlap with the gtests. So my feeling is that the IR tests could be thinned out, but some other opinions would be good.