@eme64 2025-02-13
@j3graham Can you please update the PR description at the top? The current version does not reflect the most up-to-date explanations, right?

I would like to see a nice summary, what cases were covered before and what cases you are now covering additionally.
Give a quick explanation how you changed the code.

I'll increase the number of reviewers as this looks like a substantial change.

/reviewers 2 reviewer