@eme64 2025-02-13
I also see that https://github.com/openjdk/jdk/pull/2776 and https://github.com/openjdk/jdk/pull/4136 were mentioned here. Both of those are related an have no IR tests of their own, yikes! We have to ensure that we cover those old cases, and then new ones here, so that we do not get any accidental regressions.

Maybe that's all already covered in other existing tests or the tests you added. Can you please provide a summary of all tests and what cases they cover in the PR description? It would help a lot for reviewing.