@j3graham 2025-02-13
I have updated the summary to be more informative. I believe the related optimizations are now covered in tests.