@j3graham 2025-03-24
Fixed the description and added randomized tests for Long. Happily min/max over Long is now optimized!