@liach 2025-03-26
@iwanowww Can you help review this, which causes expand and compress with constant mask not constant fold on aarch (x64 uses instructions)